refactor: organize fixtures and use collection name constant
- Create Fixtures folder in IntegrationTests project
- Move PostgresFixture to Fixtures/PostgresFixture.cs
- Extract and move PostgresCollection to Fixtures/PostgresCollection.cs
- Add Name constant to PostgresCollection
- Use PostgresCollection.Name instead of hardcoded string
- Update namespaces to Dapper.AmbientContext.IntegrationTests.Fixtures
- Better organization and maintainability
