+++ test/Dapper.AmbientContext.IntegrationTests/Fixtures/PostgresCollection.cs
@@ -1,0 +1,9 @@
+using Xunit;
+
+namespace Dapper.AmbientContext.IntegrationTests.Fixtures;
+
+[CollectionDefinition(Name)]
+public class PostgresCollection : ICollectionFixture<PostgresFixture>
+{
+    public const string Name = "Postgres";
+}
